CI: install pkg-config
This seems not to be needed, having it as a precaution in case it would
fix any future issues.
Should be re-visited for dropping this patch.

Background:
The GitHub Actions runner on Windows have Strawberry Perl,
which puts their pkg-config on the PATH.

Meson is ignoring that explicitly, as it caused issues in meson to use
Strawberry Perl's pkg-config.

While trying to fix another build issue,
I noticed that meson had a warning about this.
I'm not sure if pkg-config is needed for the build;
for now better to have a proper one than a broken one.